Format: Add an automated formater script in package.json
Till now every file had different format specially the
rendered one, it was making the whole thing wrapped
We don't need this, the browser is bound to make
that change when the code is routed to the end user,
rather it's better for us to store a formatted code
using bun run fmt which will find all the .js
.svelte and fix their format.

This will be made as a proper format further down
the line and even fixing routes directory format
as we write the master code.
Therefore

Give a fmt command to the source code.

Signed-off-by: K Jayatheerth <jayatheerthkulkarni2005@gmail.com>